Change how lua server checks spelling
The lua interface to the hunspell C API appeared to not be
able to handle some unicode chars correctly (it kept saying
Éclair was not a word, even though if you checked it on
the command line, it was).

Throw out the luaspell module, and just use a little shell
script to check the spelling and print the result so lua
could read it easily.